Hopefully final fix to get ERS working in Travis
There's no easy way to get the 'avahi' and 'dbus' python packages to
install in Travis. I've modified zeroconf.py so that it if these aren't
installed, the zeroconf package still imports normally. An exception
will only be thrown when the program actually attempts to perform any
Avahi operations (which is not the case during the test script).